:recycle: unwraps chained conditionals
with an array_map().  This improves the complexity rating, and makes the code more readable.  It does raise an alarm about too many return statements, but I think the improved readability is worth the compromise at this point.  A future refactor could get to the deeper complexity issue of the number of branching paths at work here.