:construction_worker: removing PHP 7.4 from test matrix
because travis-ci tests don't work with 7.4.  Problem appears to be a mismatch with the version of phpunit that comes with PHP on travis-ci.  I think there is an issue where travis is ignoring the phpunit being downloaded in the script.  This issue should probably also be addressed in the upstream model repo: https://github.com/wp-cli/sample-plugin